Initial port from 'Daemons' to 'Pentagram'
This commit does the initial port of Slow Query Exporter from 'Daemons' to
'Pentagram'. The move to Pentagram is intended to gain us several advantages:

- graceful signal handling.
- standardized argument handling.
- protection from running the daemon multiple times concurrently.

The chief difficulty with the initial implementation was arranging the input
(reading) of the MySQL slow query log in such a way that it was interruptible
and thus resumable. This was achieved by implementing a new #gets_with_timeout
method that can be used to read a line from the MySQL slow query log, but to
return nil if no line is available to be read before <timeout> seconds. Use of
this method allows us to implement the main daemon loop in such a way that it
is not blocked for unbounded amounts of time while it waits for further lines of
data from the slow query log.