Read chunks of data rather than lines of data
The initial implementation of #gets_with_timeout was highly inefficient in that
it read data from the file descriptor one character (one byte) at a time. This
unexpectedly caused repeated system calls. The initial expectation was that ruby
would provide read buffering for file descriptors as a matter of course, but
testing has shown that this is not the case.

Therefore we refactor things to read "chunks" of data from the slow query log
and then manually split the data into lines (and protect against a line being
split across multiple chunk reads, which should be a very common occurrence).

This is a much more efficient way to achieve the same ends that had been
originally pursued in the implementation of #gets_with_timeout.