Refactor main loop to use Pentagram::SignalBroker.select
The core issue with the previous implementation of hook_main was that it was
effectively a busy loop with an embedded sleep() call within it (effectively it
polled for queued signals). The new implementation takes advantage of the custom
`select` method provided by Pentagram::SignalBroker to treat signals as
direct interrupts to the core IO.select call. With this implementation the main
loop will spend most of it's time simply waiting for file descriptor activity
and then handling it when it occurs.

Any polling is no longer necessary, but the sleep/delay remains an option,
solely as a control for the rate at which messages will be sent to the target
GELF server.