Resolve target GELF host immediately at startup
The 'gelf' module does not attempt to resolve the target hostname until it tries
to send it data. Therefore, if we pass in an invalid hostname it could be an
arbitrary length of time until SQE 'notices'. While we can't completely get rid
of this issue, we can at least mitigate it by doing initial validation that the
hostname is valid (so if someone fat-fingers a hostname, the script will
complain immediately).